+++ fmtutil/example_test.go
@@ -60,6 +60,19 @@ func ExamplePrettyNum() {
 	// 6128750.26 → 6|128|750.26
 }
 
+func ExamplePrettyBool() {
+	fmt.Printf("%t → %s\n", true, PrettyBool(true))
+	fmt.Printf("%t → %s\n", false, PrettyBool(false))
+	fmt.Printf("%t → %s\n", true, PrettyBool(true, "Yep", "Nope"))
+	fmt.Printf("%t → %s\n", false, PrettyBool(false, "Yep", "Nope"))
+
+	// Output:
+	// true → Y
+	// false → N
+	// true → Yep
+	// false → Nope
+}
+
 func ExamplePrettyPerc() {
 	var (
 		n1 float64 = 0.123